change CurrentAvatarState.actionPoint to ReactiveAvatarState.ActionPoint
더 이상 AvatarState 모델 안에 있는 actionPoint를 직접 접근해선 안됨

+++ nekoyume/Assets/_Scripts/UI/Widget/Popup/BattleResultPopup.cs
@@ -51,7 +51,7 @@ namespace Nekoyume.UI
             public int WorldID;
             public int StageID;
             public int ClearedWaveNumber;
-            public int ActionPoint;
+            public long ActionPoint;
             public int LastClearedStageId;
             public bool ActionPointNotEnough;
             public bool IsClear;
